hecho hasta ft_memcpy

+++ libft/ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: manufern <manufern@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 17:40:43 by manufern          #+#    #+#             */
-/*   Updated: 2023/09/12 17:56:29 by manufern         ###   ########.fr       */
+/*   Updated: 2023/09/13 18:19:58 by manufern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 /* devuelve la longitud de un string */
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
